feat: opportunityBody component
